--- drivers/i2c/slave/Kconfig.eeprom
@@ -1,9 +1,0 @@
-# I2C EEPROM Slave configuration options
-
-# Copyright (c) 2017 BayLibre, SAS
-# SPDX-License-Identifier: Apache-2.0
-
-config I2C_EEPROM_SLAVE
-	bool "I2C Slave EEPROM driver"
-	help
-	  Enable virtual I2C Slave EEPROM driver
