+++ tests/drivers/i2c/i2c_slave_api/src/eeprom.h
