+++ tests/drivers/i2c/i2c_slave_api/src/main.c
@@ -17,7 +17,7 @@ LOG_MODULE_REGISTER(main);
 #include <sys/util.h>
 
 #include <drivers/i2c.h>
-#include <drivers/i2c/slave/eeprom.h>
+#include "eeprom.h"
 
 #include <ztest.h>
